backend/src/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5
chardet==3.0.4
cycler==0.10.0
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-extensions==3.0.4
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
filelock==3.0.12
graphviz==0.14.1
idna==2.9
imageio==2.6.1
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
oauthlib==3.1.0
pandas==1.0.3
Pillow==6.2.1
pipenv==2018.11.26
psycopg2==2.8.5
pydot==1.4.1
pyparsing==2.4.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rest-pandas==1.1.0
scipy==1.3.3
six==1.13.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.9
virtualenv-clone==0.5.4
xlwt==1.3.0
